GeometryOptimization

Documentation for GeometryOptimization.

SciMLBase.OptimizationFunction

GeometryOptimization.clamp_atoms

GeometryOptimization.not_clamped_mask

GeometryOptimization.update_not_clamped_positions

GeometryOptimization.update_positions

SciMLBase.OptimizationFunction — Method

By default we work in cartesian coordinaes. Note that internally, when optimizing the cartesian positions, atomic units are used.

GeometryOptimization.clamp_atoms — Method

Clamp given atoms in the system. Clamped atoms are fixed and their positions will not be optimized. The atoms to be clamped should be given as a list of indices corresponding to their positions in the system.

GeometryOptimization.not_clamped_mask — Method

Returns a mask for selecting the not clamped atoms in the system.

GeometryOptimization.update_not_clamped_positions — Method

Creates a new system based on $system$ where the non clamped positions are updated to the ones provided (in the order in which they appear in the system).

GeometryOptimization.update_positions — Method

Creates a new system based on $system$ but with atoms positions updated to the ones provided.